use backup translink APIs
