+++ README.md
@@ -1,25 +1,29 @@
 # TransLink Bot
+
 A simple Telegram chatbot to query TransLink bus schedules
 
 ## How to run?
 
 1. Create a `.env` file in the root directory with the below details
+
 ```
 BOT_TOKEN = <YOUR_TELEGRAM_BOT_TOKEN>
-TRANSLINK_API_KEY = <YOUR_TRANSLINK_API_KEY>
 ```
 
 2. Install dependencies
+
 ```
 $ npm install
 ```
 
 3. Start application
+
 ```
 $ npx nodemon src/app.ts
 ```
 
 ## How to use?
-The bot is available at [https://t.me/TransLinkBot](https://t.me/TransLinkBot)
+
+The bot is available at [https://t.me/DummyTransitBot](https://t.me/DummyTransitBot)
 
 ![image](https://user-images.githubusercontent.com/42204888/225764750-cd05728d-0cb7-4dff-8ca6-c72d9701d80c.png)
