+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "engines": {
-    "node": "18.15.0"
+    "node": "18.17.1"
   },
   "main": "dist/app.js",
   "scripts": {
@@ -18,6 +18,7 @@
     "axios": "^1.6.0",
     "dotenv": "^16.0.3",
     "lodash": "^4.17.21",
+    "moment-timezone": "^0.5.45",
     "telegraf": "^4.12.2"
   },
   "devDependencies": {
