+++ src/app.ts
@@ -1,8 +1,8 @@
-import { Context, Markup, Telegraf } from "telegraf";
-import { Location, Update } from "typegram";
+import { Context, Telegraf } from "telegraf";
+import { Update } from "typegram";
 import { message } from "telegraf/filters";
 import dotenv from "dotenv";
-import { getBusArrivalTime, getNearby, NearbyResponse } from "./schedules";
+import { getBusArrivalTime } from "./arrival";
 
 dotenv.config();
 
@@ -11,29 +11,9 @@ const bot: Telegraf<Context<Update>> = new Telegraf(
 );
 
 bot.start((ctx) => {
-  const message: string = `Welcome! Please use the command /nearby to find nearby bus stops and schedules.\n
-Or reply with the below format to query bus arrival time:
+  const message = `Welcome! Please reply with the below format to query bus arrival time:
 "<bus_id> <bus_stop_id>" (e.g. "145 59314")`;
   ctx.reply(message);
-});
-
-bot.command("nearby", async (ctx) => {
-  return ctx.reply(
-    "Please share your location",
-    Markup.keyboard([Markup.button.locationRequest("Send location")])
-      .resize(true)
-      .oneTime(true)
-  );
-});
-
-bot.on(message("location"), async (ctx) => {
-  const location: Location.CommonLocation = ctx.message.location;
-  const nearbyResponse: NearbyResponse = await getNearby(location);
-
-  ctx.reply(
-    nearbyResponse.message,
-    Markup.keyboard(nearbyResponse.options).resize(true).oneTime(false)
-  );
 });
 
 bot.on(message("text"), async (ctx) => {
